auto downloading subtitles through subliminal:
When the server is started with the --subliminal option, it will try to
download a subtitles file matching the torrent with subliminal (an
external python program).

You can also pass a --subliminal_args option when starting the server:
these are the options passed to the `subliminal download` command.
This allows you to, for example, choose the subtitles language.